fix: update token handling in BiLogsCRUD and adjust BiLogsOutputSchema to use float for tokens

+++ dynamo_lib/supabase_bi/schemas/bi_logs_schema.py
@@ -49,6 +49,6 @@ class BiLogsOutputSchema(BaseModel):
     clients_count: int = Field(
         ..., alias="clients_count", description="Count of unique clients in logs"
     )
-    tokens: Optional[int] = Field(
+    tokens: Optional[float] = Field(
         None, alias="tokens", description="Number of tokens used in the log entry"
     )
